fix hashParticipant from 64 to 32

+++ x/schat/types/message_get_conversation_key.go
@@ -9,7 +9,7 @@ const TypeMsgGetConversationKey = "get_conversation_key"
 
 var _ sdk.Msg = &MsgGetConversationKey{}
 
-func NewMsgGetConversationKey(creator string, hashParticipant []uint64) *MsgGetConversationKey {
+func NewMsgGetConversationKey(creator string, hashParticipant []uint32) *MsgGetConversationKey {
 	return &MsgGetConversationKey{
 		Creator:         creator,
 		HashParticipant: hashParticipant,
